backend/requirements.txt
torch>=2.0.0
transformers>=4.30.0
diffusers>=0.19.0
accelerate>=0.21.0
fastapi>=0.100.0
python-multipart>=0.0.6
pillow>=10.0.0
uvicorn>=0.23.0
numpy>=1.24.0
